Add BalanceCards, Charts, and TransactionSummary components for financial overview
